+++ tests/multi_period_constraint_tests.py
@@ -18,7 +18,7 @@ import pytest
 from pyomo.repn.tests.lp_diff import lp_diff
 
 from oemof import solph
-from test_scripts.test_solph.test_multi_period_model.test_multi_period_investment_model import (
+from test_scripts.test_solph.test_multi_period.test_investment_model import (
     set_up_multi_period_investment_model,
 )
