+++ tests/test_scripts/test_solph/test_multi_period/test_investment_model.py
